feat: add identifier to transport interface messages
A transport implementation can now attach an additional identifier to an envelope. This identifier can be used to faciliate multiplexing if multiple broker connections are used simultaneously. If the transport implementation does not support multiplexing, it can be left empty.
The current messenger implementation just forwards the identifier from the receiving message to the responding message.